+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Triggers/CardMovedToListTrigger.cs
@@ -44,7 +44,7 @@ namespace TrelloDotNet.AutomationEngine.Model.Triggers
         }
 
         /// <summary>
-        /// If the Trigger is met5
+        /// If the Trigger is met
         /// </summary>
         /// <remarks>
         /// While this is built to support async execution, It is best practice to keep a Trigger as light as possible only checking values against the webhook Action and not make any API Calls. The reason for this is that Triggers are called quite often, so it is better to make simple triggers and supplement with Conditions